12 Fix bug deleting entities
This commit:

* Fixes a bug in which entities deleted by the server were not deleted in state.

This happened when the server did not respond with the deleted entity’s ID, which is a likely scenario causing a bug in most cases.

To delete an entity, dispatch the `destroy` thunk action passing in the entity to delete as the parameter. The entity to delete must have an `id` key.